Update TITLE_1_CH_1.md
Deletion of ”
The government will publish an annual report detailing the measures taken to regulate the media landscape, including any restrictions applied, with clear justifications and an assessment of their proportionality and necessity. This report will be examined by an independent monitoring body to ensure transparency and accountability”, as we should instead create a single article later to oblige the government to publish annual reports on its actions (it should do so for many other things).